refactor: broken down the HTTP request into smaller chunks
